% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.r
\name{set.bibfile}
\alias{set.bibfile}
\alias{set.check.bib}
\title{Set the bibfile path}
\usage{
set.bibfile(bibfile)

set.check.bib(check)
}
\description{
Can use `options(...)` but this is laziness because I have tricksy names for
some of the options.
}
\section{Functions}{
\itemize{
\item \code{set.check.bib}: set whether bib files are validated on load and query
}}

